fix: apply review feedback on damping models and click mapper
- Remove unused DampingDirection enum (never referenced in code)
- Fix DampingType.biLinear doc formula: offset term is offset×sign(v),
  not bare offset, to correctly model both compression and rebound
- Change ClickMapper.clicksToCoefficient clicks/clicksRange params
  from int to double, matching SuspensionParameters.compression/rebound
  (double, Slider-driven) and eliminating casting at every call site
- Fix threshold branch test: use velocityMps: threshold directly and
  update comment to clarify the high-speed branch is taken at |v|==threshold
  (continuity offset ensures force still equals cLow×threshold)
- Update test loop variables from int to double to match new API

+++ lib/services/simulation/click_mapper.dart
@@ -31,9 +31,9 @@ class ClickMapper {
   /// - [clicksRange] ≤ 0, or
   /// - [baseCoefficient] ≤ 0.
   static double clicksToCoefficient(
-    int clicks,
+    double clicks,
     double baseCoefficient, {
-    int clicksRange = 20,
+    double clicksRange = 20.0,
   }) {
     if (clicksRange <= 0) {
       throw ArgumentError.value(
